+++ changelog.md
@@ -5,6 +5,7 @@
 ### 14 Sep 2016
 
 * **Fix** - Full width map no longer has 10px gutters to the left and right of the map. *DP*
+* **Fix** - Updated the preset layout, page import and duplicate page nonce actions to be more specific as not to be confused by any other nonce actions. *MP*
 * **Enhancement** - Added new Header Sidebar layout option. *MP*
 
 =======